requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yt-dlp==2025.5.22
openai-whisper==20231117
ffmpeg-python==0.2.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.0
httpx==0.25.2
streamlit==1.29.0
requests==2.31.0 